% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.first.letters.R
\name{change.first.letters}
\alias{change.first.letters}
\title{change.first.letters}
\usage{
change.first.letters(vector)
}
\arguments{
\item{Click}{here and there}
}
\description{
Changes the first letter of each word to capital letter except for help-words such as
}
\examples{
checking.nominal()
}
\keyword{nominal}
