% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{find_files}
\usage{
find_files(lookfor, p = "C:/Users/", filetype = "R", alltolower = TRUE,
  subdir = TRUE)
}
\arguments{
\item{lookfor}{string to be found}

\item{p}{Path where the search takes place. The more precise, the shorter the time to find. Default is "C:/Users/"}

\item{filetype}{type of file. Default is "R"}

\item{alltolower}{boolean. casesenstivity. default is TRUE}

\item{subdir}{should the search be extended to subdirectories. Default is TRUE}
}
\description{
find_files
}
\examples{

}
