% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.grep.vector.R
\name{get.grep.vector}
\alias{get.grep.vector}
\title{get.grep.vector}
\usage{
get.grep.vector(vec, x)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
