% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_identify.scriptstyle.section.R
\name{html_identify.scriptstyle.section}
\alias{html_identify.scriptstyle.section}
\title{html_identify.scriptstyle.section}
\usage{
html_identify.scriptstyle.section(x)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
