% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_remove.emphased.R
\name{html_remove.emphased}
\alias{html_remove.emphased}
\title{html_remove.emphased}
\usage{
html_remove.emphased(x)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
