% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigrepl.R
\name{multigrepl}
\alias{multigrepl}
\title{multigrepl}
\usage{
multigrepl(vector, greplvector, ...)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
