% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.colors.R
\name{random.colors}
\alias{random.colors}
\title{random.colors}
\usage{
random.colors(nmbr, seed = 1337, exclude_color = "",
  exclude.vector = NULL, distinctiveness = 60, removetoobright = TRUE)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\keyword{nominal}
