% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.colors_old.R
\name{random.colors_old}
\alias{random.colors_old}
\title{random.colors}
\usage{
random.colors_old(nmbr, seed = 1337, exclude_color = "",
  exclude.vector = NULL)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\keyword{nominal}
