% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_timestamp.R
\name{read_csv_timestamp}
\alias{read_csv_timestamp}
\title{read_csv_timestamp function}
\usage{
read_csv_timestamp(filepath, filename)
}
\arguments{
\item{filepath}{path to file}

\item{filename}{name of the file}
}
\description{
read_csv_timestamp function
}
