% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_version.R
\name{read_csv_version}
\alias{read_csv_version}
\title{read_csv_version function}
\usage{
read_csv_version(filepath, filename)
}
\arguments{
\item{filepath}{path to file}

\item{filename}{name of file}
}
\description{
Reads versioned file. Corresponding to write_csv_version
}
