% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKML.R
\name{readKML}
\alias{readKML}
\title{readKML}
\usage{
readKML(file, keep_name_description = FALSE, ...)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\keyword{nominal}
