% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.columns.R
\name{rearrange.columns}
\alias{rearrange.columns}
\title{rearrange.columns}
\usage{
rearrange.columns(df, pos, var)
}
\arguments{
\item{Click}{here and there}
}
\description{
Rearrange position of variables
}
\examples{
checking.nominal()
}
\keyword{nominal}
