% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{repath}
\alias{repath}
\title{repath}
\usage{
repath()
}
\description{
Changes backslashes in string in clipboard when run. Useful for windows paths.
}
\note{
Code heavily inspired by https://stackoverflow.com/questions/1189759/expert-r-users-whats-in-your-rprofile/12703931#12703931.
}
\examples{


}
