% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.whitespace.R
\name{text_clean.whitespace}
\alias{text_clean.whitespace}
\title{text_clean.whitespace}
\usage{
text_clean.whitespace(vec)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
