% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_rem.special.letters.R
\name{text_rem.special.letters}
\alias{text_rem.special.letters}
\title{text_rem.special.letters}
\usage{
text_rem.special.letters(x)
}
\arguments{
\item{Click}{here and there}
}
\description{
check nominal if they are nominal
}
\examples{
checking.nominal()
}
\keyword{nominal}
