% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_counter.R
\name{version_counter}
\alias{version_counter}
\title{version_counter function}
\usage{
version_counter(filepath, filename, filetype = ".Rmd", increaseversion = 1)
}
\arguments{
\item{filepath}{path to file}

\item{filename}{name to file}

\item{filetype}{type of fileending. Default is ".Rmd"}

\item{increaseversion}{by which number should the version counter be increased? Default is 1.}
}
\description{
check ordinal if they are ordinal
}
\examples{

}
