% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_timestamp.R
\name{write_csv_timestamp}
\alias{write_csv_timestamp}
\title{write_csv_timestamp function}
\usage{
write_csv_timestamp(data, filepath, filename)
}
\arguments{
\item{Click}{here and there}
}
\description{
check ordinal if they are ordinal
}
\examples{
get.1.sub(data,split,cols,whichcol,replace)
}
\keyword{ordinal}
