% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{find_files}
\usage{
find_files(lookfor, p = "C:/Users/", filetype = "R", alltolower = TRUE,
  subdir = TRUE)
}
\arguments{
\item{Click}{here and there}
}
\description{
find_files
}
\examples{

}
\keyword{nominal}
