% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_version.R
\name{read_csv_version}
\alias{read_csv_version}
\title{read_csv_version function}
\usage{
read_csv_version(filepath, filename)
}
\arguments{
\item{Click}{here and there}
}
\description{
check ordinal if they are ordinal
}
\examples{
get.1.sub(data,split,cols,whichcol,replace)

}
\keyword{ordinal}
