% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.whitespace.R
\name{text_clean.whitespace}
\alias{text_clean.whitespace}
\title{text_clean.whitespace}
\usage{
text_clean.whitespace(vec)
}
\arguments{
\item{vec}{character string or vector of character strings}
}
\description{
Cleans double whitespaces and whitespace at the beginning or end of string.
}
\examples{
test<-c("te    st"," test ")
test
text_clean.whitespace(test)
}
