% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_counter.R
\name{version_counter}
\alias{version_counter}
\title{version_counter function}
\usage{
version_counter(filepath, filename, filetype = ".Rmd", increaseversion = 1)
}
\arguments{
\item{Click}{here and there}
}
\description{
check ordinal if they are ordinal
}
\examples{
get.1.sub(data,split,cols,whichcol,replace)


}
\keyword{ordinal}
